// src/interfaces/Interface.ts
export interface Profile {
    fullName: string | undefined;
    birhtDate: string | undefined;
    avatar: string | undefined;
}